requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mutagen==1.47.0
google-generativeai==0.3.2
aiofiles==23.2.1
watchdog==4.0.0
sqlalchemy==2.0.25
python-multipart==0.0.6
sse-starlette==1.8.2
